removed PTCWallet
